+++ src/main/java/com/renault/garagemiscroservice/repositories/GarageRepository.java
@@ -3,5 +3,5 @@ package com.renault.garagemiscroservice.repositories;
 import com.renault.garagemiscroservice.entities.Garage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GarageRepository extends JpaRepository<Garage,Long> {
+public interface GarageRepository extends JpaRepository<Garage,Integer> {
 }
