+++ src/main/java/com/renault/garagemiscroservice/services/AccessoireService.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface AccessoireService {
     void createAccessoire(AccessoireDTO accessoireDTO) throws EntityNotFoundException, MethodArgumentNotValidException;
     void updateAccessoire(AccessoireDTO accessoireDTO) throws MethodArgumentNotValidException, EntityNotFoundException;
-    void deleteAccessoire(Long id) throws MethodArgumentNotValidException, EntityNotFoundException;
-    List<AccessoireDTO> getAllAccessoiresByVehicule(Long id) throws MethodArgumentNotValidException;
+    void deleteAccessoire(Integer id) throws MethodArgumentNotValidException, EntityNotFoundException;
+    List<AccessoireDTO> getAllAccessoiresByVehicule(Integer id) throws MethodArgumentNotValidException;
 
 }
